+++ packages/eyes-sdk-core/test/utils/MockDriver.js
@@ -1,7 +1,13 @@
 const png = require('png-async')
+const {inspect} = require('util')
+const snippets = require('@applitools/snippets')
 const {TypeUtils} = require('../../index')
 const FakeDomSnapshot = require('./FakeDomSnapshot')
-const snippets = require('@applitools/snippets')
+
+const WELL_KNOWN_SCRIPTS = {
+  'dom-snapshot': script => /^\/\* @applitools\/dom-snapshot@[\d.]+ \*\//.test(script),
+  'dom-capture': script => /^\/\* @applitools\/dom-capture@[\d.]+ \*\//.test(script),
+}
 
 const DEFAULT_STYLES = {
   'border-left-width': '0px',
@@ -14,8 +20,6 @@ const DEFAULT_PROPS = {
   clientHeight: 400,
   overflow: null,
 }
-
-const DOM_SNAPSHOT_REGEX = /^\/\* @applitools\/dom-snapshot@[\d.]+ \*\//
 
 class MockDriver {
   constructor({isNative = false, isMobile = false} = {}) {
@@ -31,6 +35,11 @@ class MockDriver {
     this._contexts = new Map()
     this._contexts.set(null, {
       document: {id: Symbol('documentId')},
+      state: {
+        get name() {
+          return null
+        },
+      },
     })
     this._contextId = null
     this.mockElement('html', {scrollPosition: {x: 0, y: 0}})
@@ -40,35 +49,39 @@ class MockDriver {
       const isCORS = !isRoot && context.isCORS
       const contentDocument = await this.findElement('html')
       const selector = !isCORS && !isRoot ? context.element.selector : null
-      return {isRoot, isCORS, contentDocument, selector}
+      return [contentDocument, selector, isRoot, isCORS]
     })
     this.mockScript(snippets.getChildFramesInfo, () => {
       return Array.from(this._contexts.values())
         .filter(frame => frame.parentId === this._contextId)
-        .map(frame => ({isCORS: frame.isCORS, element: frame.element}))
+        .map(frame => [frame.element, frame.isCORS])
     })
-    this.mockScript(snippets.getElementRect, ({element}) => {
+    this.mockScript(snippets.getElementRect, ([element]) => {
       return element.rect || {x: 0, y: 0, width: 100, height: 100}
     })
-    this.mockScript(snippets.getElementComputedStyleProperties, ({element, properties}) => {
+    this.mockScript(snippets.getElementComputedStyleProperties, ([element, properties]) => {
       return properties.map(
         property => (element.styles || {})[property] || DEFAULT_STYLES[property],
       )
     })
-    this.mockScript(snippets.getElementProperties, ({element, properties}) => {
+    this.mockScript(snippets.getElementProperties, ([element, properties]) => {
       return properties.map(property => (element.props || {})[property] || DEFAULT_PROPS[property])
     })
-    this.mockScript(snippets.setElementStyleProperties, ({element, properties}) => {
+    this.mockScript(snippets.setElementStyleProperties, ([element, properties]) => {
       return Object.entries(properties).reduce((original, [name, value]) => {
         original[name] = element.style[name]
         element.style[name] = value
         return original
       }, {})
     })
-    this.mockScript(snippets.setElementAttribute, ({element, attr, value}) => {
-      element.attrs[attr] = value
+    this.mockScript(snippets.setElementAttributes, ([element, attributes]) => {
+      return Object.entries(attributes).reduce((original, [name, value]) => {
+        original[name] = element.attrs[name]
+        element.attrs[name] = value
+        return original
+      }, {})
     })
-    this.mockScript(snippets.scrollTo, ({element, offset}) => {
+    this.mockScript(snippets.scrollTo, ([element, offset]) => {
       let scrollingElement = element
       if (!element) {
         scrollingElement = this.findElement('html')
@@ -76,7 +89,7 @@ class MockDriver {
       scrollingElement.scrollPosition = offset
       return [scrollingElement.scrollPosition.x, scrollingElement.scrollPosition.y]
     })
-    this.mockScript(snippets.getElementScrollOffset, ({element}) => {
+    this.mockScript(snippets.getElementScrollOffset, ([element]) => {
       let scrollingElement = element
       if (!element) {
         scrollingElement = this.findElement('html')
@@ -86,7 +99,7 @@ class MockDriver {
       }
       return {x: scrollingElement.scrollPosition.x, y: scrollingElement.scrollPosition.y}
     })
-    this.mockScript(snippets.getElementInnerOffset, ({element}) => {
+    this.mockScript(snippets.getElementInnerOffset, ([element]) => {
       let scrollingElement = element
       if (!element) {
         scrollingElement = this.findElement('html')
@@ -105,7 +118,7 @@ class MockDriver {
     this.mockScript(snippets.getViewportSize, () => {
       return {width: this._window.rect.width, height: this._window.rect.height}
     })
-    this.mockScript(snippets.getElementXpath, ({element}) => {
+    this.mockScript(snippets.getElementXpath, ([element]) => {
       if (element.xpath) return element.xpath
       const elements = Array.from(this._elements.values()).reduce(
         (elements, array) => elements.concat(array),
@@ -119,14 +132,13 @@ class MockDriver {
     this.mockScript(snippets.blurElement, () => {
       return null
     })
-    this.mockScript(DOM_SNAPSHOT_REGEX, () => FakeDomSnapshot.generateDomSnapshot(this))
-    this.mockScript(snippets.markElements, ({elements, ids}) => {
+    this.mockScript(snippets.markElements, ([elements, ids]) => {
       for (const [index, el] of elements.entries()) {
         el.attributes = el.attributes || []
         el.attributes.push({name: 'data-eyes-selector', value: ids[index]})
       }
     })
-    this.mockScript(snippets.cleanupElementIds, ({elements}) => {
+    this.mockScript(snippets.cleanupElementIds, ([elements]) => {
       for (const el of elements) {
         el.attributes.splice(
           el.attributes.findIndex(({name}) => name === 'data-eyes-selector'),
@@ -134,9 +146,7 @@ class MockDriver {
         )
       }
     })
-  }
-  static get DOM_SNAPSHOT_REGEX() {
-    return DOM_SNAPSHOT_REGEX
+    this.mockScript('dom-snapshot', () => FakeDomSnapshot.generateDomSnapshot(this))
   }
   mockScript(scriptMatcher, resultGenerator) {
     this._scripts.set(scriptMatcher, resultGenerator)
@@ -165,6 +175,11 @@ class MockDriver {
         isCORS: state.isCORS,
         element,
         document: {id: Symbol('documentId' + Math.floor(Math.random() * 100))},
+        state: {
+          get name() {
+            return element.selector
+          },
+        },
       })
       element.contextId = contextId
       this.mockElement('html', {
@@ -189,16 +204,15 @@ class MockDriver {
   }
   async executeScript(script, args = []) {
     args = serialize(args)
-    let resultGenerator = this._scripts.get(script)
-    if (!resultGenerator) {
-      for (const [tester, result] of this._scripts.entries()) {
-        if (TypeUtils.isFunction(tester.test) && tester.test(script)) {
-          resultGenerator = result
-          break
-        }
-      }
+    let result = this._scripts.get(script)
+    if (!result) {
+      const name = Object.keys(WELL_KNOWN_SCRIPTS).find(name => WELL_KNOWN_SCRIPTS[name](script))
+      console.log
+      if (!name) return null
+      result = this._scripts.get(name)
     }
-    return TypeUtils.isFunction(resultGenerator) ? resultGenerator(...args) : resultGenerator
+    const {state} = this._contexts.get(this._contextId)
+    return TypeUtils.isFunction(result) ? result.call(state, ...args) : result
   }
   async findElement(selector) {
     const elements = this._elements.get(selector)
@@ -268,13 +282,13 @@ class MockDriver {
     })
   }
   toString() {
-    return 'MockDriver'
+    return '<MockDriver>'
   }
   toJSON() {
-    return 'MockDriver'
+    return '<MockDriver>'
   }
-  [require('util').inspect.custom]() {
-    return 'MockDriver'
+  [inspect.custom]() {
+    return '<MockDriver>'
   }
 }