+++ packages/snippets/src/browser/getChildFramesInfo.js
@@ -1,10 +1,10 @@
 function getChildFramesInfo() {
   const frames = document.querySelectorAll('frame, iframe')
-  return Array.prototype.map.call(frames, frameElement => ({
-    isCORS: !frameElement.contentDocument,
-    element: frameElement,
-    src: frameElement.src,
-  }))
+  return Array.prototype.map.call(frames, frameElement => [
+    frameElement,
+    !frameElement.contentDocument,
+    frameElement.src,
+  ])
 }
 
 module.exports = getChildFramesInfo