+++ packages/snippets/src/browser/setElementStyleProperties.js
@@ -1,4 +1,4 @@
-function setElementStyleProperties({element, properties} = {}) {
+function setElementStyleProperties([element, properties] = []) {
   return Object.keys(properties).reduce((original, property) => {
     original[property] = element.style[property]
     element.style[property] = properties[property]